% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{check_for_updates}
\alias{check_for_updates}
\title{check_for_updates}
\usage{
check_for_updates(line_nb, envi = .GlobalEnv)
}
\arguments{
\item{line_nb}{The line number of the data to be checked for updates on the data_default reference table.}

\item{envi}{The environment in which to look for the reference table.}
}
\value{
The date the dataset was last updated in Date formate. (For APIs always returns the current date + 1 day)
}
\description{
Check for updates. Returns true if needs updating, can take dataset name as param.
}
