% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{condition}
\alias{condition}
\title{condition}
\usage{
condition(subclass, message, call = sys.call(-1), ...)
}
\arguments{
\item{subclass}{Description The class of the new condition.}

\item{message}{The error message to be shown to the user.}

\item{call}{???}
}
\value{
A custom condition that can be used
}
\description{
Creates a custom condition to be raised during errors.
}
\examples{
condition("customErrorName", "Custom message")
}
\author{
Hadley Wickham
}
