% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\docType{data}
\name{data_default}
\alias{data_default}
\title{data_default}
\format{
An object of class \code{data.frame} with 3 rows and 15 columns.
}
\usage{
data_default
}
\description{
The reference data.frame used in some of the functions. Contains relevant information to automate data prepossessing in order to
simplify data management in a shiny application.
}
\keyword{datasets}
