% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\docType{data}
\name{data_sets}
\alias{data_sets}
\title{data_sets}
\format{
An object of class \code{data.frame} with 1 rows and 5 columns.
}
\usage{
data_sets
}
\description{
The table containing the list of all data sets compiled from the servers base datasets
}
\keyword{datasets}
