% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{format_col_vect}
\alias{format_col_vect}
\title{format_col_vect}
\usage{
format_col_vect(
  data,
  var_name = deparse(substitute(data)),
  vector,
  envir = .GlobalEnv,
  ...
)
}
\arguments{
\item{data}{The data to be formatted.}

\item{var_name}{The name of the variable to be assigned to the environment. Default is set to the name of the inputed data.}

\item{vector}{The vector indicating the type the columns will be formatted to. 1 = numeric, 2 = character, 3 = Date, 4 = factor.}

\item{envir}{The environment in which the outputed data is to be placed. Default is set to \code{.GlobalEnv}.}

\item{...}{The parameters to ba passed to the factor function.}
}
\value{
Nothing. Assigns the data directly to the specified environment.
}
\description{
Will set the data types to data types found in given vector. (!!! The data frame to be changed must already be present in the given environment !!!)
}
