% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{format_date_scrape}
\alias{format_date_scrape}
\title{format_date_scrape}
\usage{
format_date_scrape(text, sep = "default")
}
\arguments{
\item{text}{The text to be formated as a date}

\item{sep}{The chacter used to seperate the different elements of the date. (Not yet implemented)}
}
\value{
A usable date.
}
\description{
Formats dates from raw untidy string into usable date format.
Can format dates with months as text.
(Function designed to have more features)
}
\examples{
out = format_date_scrape("#      07 juillet 2023    #")
print(out)
# 2023-07-07
class(out)
# "Date"
}
