% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_date_head}
\alias{get_date_head}
\title{get_date_head}
\usage{
get_date_head(url)
}
\arguments{
\item{url}{The url of the site the date is to be retreaved from}
}
\value{
A usable date.
}
\description{
Retrieves the date from httr request header.
Header must have a `last-modified` tag.
}
