% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_date_scrape}
\alias{get_date_scrape}
\title{get_date_scrape}
\usage{
get_date_scrape(url, selector)
}
\arguments{
\item{url}{The url of the site the date is to be retreaved from}

\item{selector}{The CSS selector used to identefy the date on the given page}
}
\value{
A usable date.
}
\description{
Scrapes the given web page for the contense of the element identified by the CSS selector.
Formates the result as a date.
}
