% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_INSEE_Commune_Data_API}
\alias{get_INSEE_Commune_Data_API}
\title{get_INSEE_Commune_Data_API}
\usage{
get_INSEE_Commune_Data_API(token = "")
}
\arguments{
\item{token}{Token used to access the INSEE API. !!! Not yet implemented function will automatically call the \code{get_INSEE_token}
function. Will be used if future versions with stored token to reduce internet trafic. !!!}
}
\value{
List containing in [[1]] the user message returned by the API and in [[2]] the data.
}
\description{
Downloads the commune data from INSEE using there API.
Returns the message to be shown to the user at [[1]] and the dataset at [[2]]
Throws error_failed_connection error if can't connect to website
}
