% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_url_file.gz}
\alias{get_url_file.gz}
\title{get_url_file.gz}
\usage{
get_url_file.gz(url, csv_type = "default")
}
\arguments{
\item{url}{Description The class of the new condition.}

\item{csv_type}{The csv function to be used to read the file. Defaults to "default".}
}
\value{
The file outputed by the read.csv functions.
}
\description{
Fetches the gz compressed file at specified url.
}
\examples{
get_url_file.gz("https::www.data_url.com", csv_url = "csv2")
get_url_file.gz("https::www.data_url.com", csv_url = "csv")
}
