% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{is.gzfile}
\alias{is.gzfile}
\title{is.gzfile}
\usage{
is.gzfile(file)
}
\arguments{
\item{file}{An object of file class.}
}
\value{
Boolean indicating if file is a gzfile.
}
\description{
Tests to see if the given file type is a gzfile.
}
