% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{is_in_string}
\alias{is_in_string}
\title{is_in_string}
\usage{
is_in_string(str, patern, ignore.case = TRUE)
}
\arguments{
\item{str}{The string in which to look for the patern}

\item{patern}{The patern to be found within the string}

\item{ignore.case}{Boolean value indicating if patern recognition should be case sensitive}
}
\value{
The boolean value indicating
}
\description{
Returns true if the patern is in the string, else false.
}
\examples{
if(is_in_string("Hello","ll",ignore.case = TRUE)){
  print("Yes it's in the string")
}
}
