% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{merge_data_set}
\alias{merge_data_set}
\title{merge_data_set}
\usage{
merge_data_set(line_nb, envi = .GlobalEnv)
}
\arguments{
\item{line_nb}{The line number of the data to be checked for updates on the data_default reference table.}

\item{envi}{The environment in which the data_default reference table is present and in which the outputed data will be place. (Yes I know. And no I don't know what I was thinking when I wrote this).}
}
\value{
Nothing. Assigns the data directly to the specified environment.
}
\description{
Given the line number in the data_sets table will reconstruct the data set in question from it's constituent data sets and assign it to it's variable name
all within
}
