% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{multi_string_input}
\alias{multi_string_input}
\title{trim_string}
\usage{
multi_string_input(str)
}
\arguments{
\item{str}{The input string to be split into vector of substrings.}
}
\value{
A vector of strings resulting from the split of the input string.
}
\description{
Function used to split single string search inputs into appropriate input for data selection function.
}
