% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{select_set_fr_crime_commune}
\alias{select_set_fr_crime_commune}
\title{select_set_fr_crime_commune}
\usage{
select_set_fr_crime_commune(
  data,
  metric = NULL,
  ndiff = FALSE,
  annee = NULL,
  classe = NULL,
  faits = NULL,
  tauxpourmille = NULL,
  pop = NULL,
  code_postal = NULL,
  nom_commune = NULL,
  code_departement = NULL,
  nom_departement = NULL,
  code_region = NULL,
  nom_region = NULL,
  use_dom_tom = FALSE
)
}
\arguments{
\item{data}{The data on which the ploting will be based. !!! Will only take the formatted data that comes with the package. !!!}

\item{metric}{By what variable the number fo counts is to be calculated: vector of column names in char format.}

\item{ndiff}{Boolean indicates if the communes with non disclosed figures should be present in the outputted data. !!! Should generally be set to FALSE. !!}

\item{annee}{Vector of the last 2 digits of the year in character format.}

\item{classe}{Vector containing the type of infraction in character format.}

\item{faits}{Vector of length 2 indicating the \code{c(minimum, maximum)} number of infraction counts to be included in the outputted data.}

\item{tauxpourmille}{Not prperly implemented yet. Leave as NULL.}

\item{pop}{Vector of length 2 indicating the \code{c(minimum, maximum)} population counts of communes to be included in the outputted data.}

\item{code_postal}{Vector of postal codes that is to be included in the outputted data.}

\item{nom_commune}{Vector of commune names that is to be included in the outputted data.}

\item{code_departement}{Vector of department codes that is to be included in the outputted data.}

\item{nom_departement}{Vector of department names that is to be included in the outputted data.}

\item{code_region}{Vector of region codes that is to be included in the outputted data.}

\item{nom_region}{Vector of region names that is to be included in the outputted data.}

\item{use_dom_tom}{Boolean value indicating if the "département et territoire d'outre mer" are to be included in the outputted data.}
}
\value{
Nothing. Assigns the data directly to the specified environment.
}
\description{
Selects the data.frame to be used for ploting selects data.
Takes vector of selected modalities as parameters.
Use NULL as the default select all as it is the value returned by selecting widget when nothing is selected.
This function was designed to be able to be used for any data representation of the set_fr_crime_commune dataset, not just the leflet map.
This function was designed to be used with a shiny interface.
}
