% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{trim_string}
\alias{trim_string}
\title{trim_string}
\usage{
trim_string(text)
}
\arguments{
\item{text}{The text to be trimed}
}
\value{
The file outputed by the read.csv functions
}
\description{
Trimes both sides of the given string of all characters that are not numbers or letters. Stops trimming once it reaches the first occurrence
of a number or letter.
}
\examples{
out = trim_string("(  a (--_Hello\"(-(-'-( '-))   )))")
print(out)
# "a (--_Hello"
}
