% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{update_all_data}
\alias{update_all_data}
\title{update_all_data}
\usage{
update_all_data(session_env = .GlobalEnv, server_env = .GlobalEnv)
}
\arguments{
\item{session_env}{The user session environment.}

\item{server_env}{The server environment.}
}
\value{
Vector of user messages informing the user on the status of the data update in character format.
}
\description{
Updates all data in the data_default table using the \code{update_data} function.
}
