% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{update_data}
\alias{update_data}
\title{update_data}
\usage{
update_data(line_nb, session_env = .GlobalEnv, server_env = .GlobalEnv)
}
\arguments{
\item{line_nb}{The line number of the data to be checked for updates on the data_default reference table.}

\item{session_env}{The user session environment.}

\item{server_env}{The server environment.}
}
\value{
The user message informing the user on the status of the data update in character format.
}
\description{
Check default_data reference table for updates, can take dataset name as parameter.
If data already up to date on server will load server data to user session.
Returns the message to inform user of download status.
Throws an error if error different than not able to connect to website.
}
